refactor: (WIP) Extract internal data classes to separate files for better organization and maintainability within the FIR transformer module.
