+++ compiler/suspend-transform-plugin/src/main/kotlin/love/forte/plugin/suspendtrans/fir/CopiedTypeParameterPair.kt
@@ -1,0 +1,8 @@
+package love.forte.plugin.suspendtrans.fir
+
+import org.jetbrains.kotlin.fir.declarations.FirTypeParameter
+
+internal data class CopiedTypeParameterPair(
+    val original: FirTypeParameter,
+    val copied: FirTypeParameter
+)
