+++ compiler/suspend-transform-plugin/src/main/kotlin/love/forte/plugin/suspendtrans/fir/CopyAnnotations.kt
@@ -1,0 +1,9 @@
+package love.forte.plugin.suspendtrans.fir
+
+import org.jetbrains.kotlin.fir.expressions.FirAnnotation
+
+internal data class CopyAnnotations(
+    val functionAnnotations: List<FirAnnotation>,
+    val propertyAnnotations: List<FirAnnotation>,
+    val toOriginalAnnotations: List<FirAnnotation>
+)
