+++ compiler/suspend-transform-plugin/src/main/kotlin/love/forte/plugin/suspendtrans/fir/FirCacheKey.kt
@@ -1,0 +1,9 @@
+package love.forte.plugin.suspendtrans.fir
+
+import org.jetbrains.kotlin.fir.scopes.impl.FirClassDeclaredMemberScope
+import org.jetbrains.kotlin.fir.symbols.impl.FirClassSymbol
+
+internal data class FirCacheKey(
+    val classSymbol: FirClassSymbol<*>,
+    val memberScope: FirClassDeclaredMemberScope?
+)
